Minor refactoring so that each query has their own file
