+++ datajunction-server/datajunction_server/api/graphql/main.py
@@ -1,23 +1,23 @@
 """DJ graphql"""
-from typing import Annotated, List, Optional
-
 import strawberry
 from fastapi import Depends
 from strawberry.fastapi import GraphQLRouter
-from strawberry.types import Info
 
-from datajunction_server.api.graphql.catalogs import CatalogInfo, list_catalogs
-from datajunction_server.api.graphql.engines import EngineInfo, list_engines
+from datajunction_server.api.graphql.queries.catalogs import list_catalogs
+from datajunction_server.api.graphql.queries.dag import common_dimensions
+from datajunction_server.api.graphql.queries.engines import list_engines
+from datajunction_server.api.graphql.queries.nodes import (
+    find_nodes,
+    find_nodes_paginated,
+)
 from datajunction_server.api.graphql.queries.sql import measures_sql
 from datajunction_server.api.graphql.queries.tags import list_tag_types, list_tags
-from datajunction_server.api.graphql.resolvers.nodes import find_nodes_by
 from datajunction_server.api.graphql.scalars import Connection
+from datajunction_server.api.graphql.scalars.catalog_engine import Catalog, Engine
 from datajunction_server.api.graphql.scalars.node import DimensionAttribute, Node
 from datajunction_server.api.graphql.scalars.sql import GeneratedSQL
 from datajunction_server.api.graphql.scalars.tag import Tag
-from datajunction_server.models.node import NodeCursor, NodeType
-from datajunction_server.sql.dag import get_common_dimensions
-from datajunction_server.utils import SEPARATOR, get_session, get_settings
+from datajunction_server.utils import get_session, get_settings
 
 
 async def get_context(
@@ -36,15 +36,36 @@ class Query:  # pylint: disable=R0903
     Parent of all DJ graphql queries
     """
 
-    list_catalogs: List[CatalogInfo] = strawberry.field(  # noqa: F811
+    # Catalog and engine queries
+    list_catalogs: list[Catalog] = strawberry.field(  # noqa: F811
         resolver=list_catalogs,
     )
-    list_engines: List[EngineInfo] = strawberry.field(  # noqa: F811
+    list_engines: list[Engine] = strawberry.field(  # noqa: F811
         resolver=list_engines,
     )
-    measures_sql: List[GeneratedSQL] = strawberry.field(  # noqa: F811
+
+    # Node search queries
+    find_nodes: list[Node] = strawberry.field(  # noqa: F811
+        resolver=find_nodes,
+        description="Find nodes based on the search parameters.",
+    )
+    find_nodes_paginated: Connection[Node] = strawberry.field(  # noqa: F811
+        resolver=find_nodes_paginated,
+        description="Find nodes based on the search parameters with pagination",
+    )
+
+    # DAG queries
+    common_dimensions: list[DimensionAttribute] = strawberry.field(  # noqa: F811
+        resolver=common_dimensions,
+        description="Get common dimensions for one or more nodes",
+    )
+
+    # Generate SQL queries
+    measures_sql: list[GeneratedSQL] = strawberry.field(  # noqa: F811
         resolver=measures_sql,
     )
+
+    # Tags queries
     list_tags: list[Tag] = strawberry.field(  # noqa: F811
         resolver=list_tags,
         description="Find DJ node tags based on the search parameters.",
@@ -53,148 +74,6 @@ class Query:  # pylint: disable=R0903
         resolver=list_tag_types,
         description="List all DJ node tag types",
     )
-
-    @strawberry.field(description="Find nodes based on the search parameters.")
-    async def find_nodes(
-        self,
-        fragment: Annotated[
-            Optional[str],
-            strawberry.argument(
-                description="A fragment of a node name to search for",
-            ),
-        ] = None,
-        names: Annotated[
-            Optional[List[str]],
-            strawberry.argument(
-                description="Filter to nodes with these names",
-            ),
-        ] = None,
-        node_types: Annotated[
-            Optional[List[NodeType]],
-            strawberry.argument(
-                description="Filter nodes to these node types",
-            ),
-        ] = None,
-        tags: Annotated[
-            Optional[List[str]],
-            strawberry.argument(
-                description="Filter to nodes tagged with these tags",
-            ),
-        ] = None,
-        *,
-        info: Info,
-    ) -> List[Node]:
-        """
-        Find nodes based on the search parameters.
-        """
-        return await find_nodes_by(info, names, fragment, node_types, tags)  # type: ignore
-
-    @strawberry.field(
-        description="Find nodes based on the search parameters with pagination",
-    )
-    async def find_nodes_paginated(
-        self,
-        fragment: Annotated[
-            Optional[str],
-            strawberry.argument(
-                description="A fragment of a node name to search for",
-            ),
-        ] = None,
-        names: Annotated[
-            Optional[List[str]],
-            strawberry.argument(
-                description="Filter to nodes with these names",
-            ),
-        ] = None,
-        node_types: Annotated[
-            Optional[List[NodeType]],
-            strawberry.argument(
-                description="Filter nodes to these node types",
-            ),
-        ] = None,
-        tags: Annotated[
-            Optional[List[str]],
-            strawberry.argument(
-                description="Filter to nodes tagged with these tags",
-            ),
-        ] = None,
-        edited_by: Annotated[
-            Optional[str],
-            strawberry.argument(
-                description="Filter to nodes edited by this user",
-            ),
-        ] = None,
-        namespace: Annotated[
-            Optional[str],
-            strawberry.argument(
-                description="Filter to nodes in this namespace",
-            ),
-        ] = None,
-        after: Optional[str] = None,
-        before: Optional[str] = None,
-        limit: Annotated[
-            Optional[int],
-            strawberry.argument(description="Limit nodes"),
-        ] = 100,
-        *,
-        info: Info,
-    ) -> Connection[Node]:
-        """
-        Find nodes based on the search parameters.
-        """
-        if not limit or limit < 0:
-            limit = 100
-        nodes_list = await find_nodes_by(
-            info,
-            names,
-            fragment,
-            node_types,
-            tags,
-            edited_by,
-            namespace,
-            limit + 1,
-            before,
-            after,
-        )
-        return Connection.from_list(
-            items=nodes_list,
-            before=before,
-            after=after,
-            limit=limit,
-            encode_cursor=lambda dj_node: NodeCursor(
-                created_at=dj_node.created_at,
-                id=dj_node.id,
-            ),
-        )
-
-    @strawberry.field(
-        description="Get common dimensions for one or more nodes",
-    )
-    async def common_dimensions(
-        self,
-        nodes: Annotated[
-            Optional[List[str]],
-            strawberry.argument(
-                description="A list of nodes to find common dimensions for",
-            ),
-        ] = None,
-        *,
-        info: Info,
-    ) -> list[DimensionAttribute]:
-        """
-        Return a list of common dimensions for a set of nodes.
-        """
-        nodes = await find_nodes_by(info, nodes)  # type: ignore
-        dimensions = await get_common_dimensions(info.context["session"], nodes)  # type: ignore
-        return [
-            DimensionAttribute(  # type: ignore
-                name=dim.name,
-                attribute=dim.name.split(SEPARATOR)[-1],
-                properties=dim.properties,
-                type=dim.type,
-            )
-            for dim in dimensions
-        ]
 
 
 schema = strawberry.Schema(query=Query)
