+++ datajunction-server/datajunction_server/api/graphql/scalars/catalog_engine.py
@@ -1,31 +1,45 @@
 """Catalog/engine related scalars"""
 
-from typing import List, Optional
-
 import strawberry
 
+from datajunction_server.models.catalog import CatalogInfo
+from datajunction_server.models.engine import EngineInfo
 from datajunction_server.models.node import Dialect as Dialect_
 
 Dialect = strawberry.enum(Dialect_)
 
 
-@strawberry.type
+@strawberry.experimental.pydantic.type(model=EngineInfo, all_fields=True)
 class Engine:  # pylint: disable=too-few-public-methods
     """
     Database engine
     """
 
-    name: str
-    version: str
-    uri: Optional[str]
-    dialect: Optional[Dialect]  # type: ignore
+
+# @strawberry.type
+# class Engine:  # pylint: disable=too-few-public-methods
+#     """
+#     Database engine
+#     """
+
+#     name: str
+#     version: str
+#     uri: Optional[str]
+#     dialect: Optional[Dialect]  # type: ignore
 
 
-@strawberry.type
+# @strawberry.type
+# class Catalog:  # pylint: disable=too-few-public-methods
+#     """
+#     Catalog
+#     """
+
+#     name: str
+#     engines: Optional[List[Engine]]
+
+
+@strawberry.experimental.pydantic.type(model=CatalogInfo, all_fields=True)
 class Catalog:  # pylint: disable=too-few-public-methods
     """
-    Catalog
+    Class for a Catalog
     """
-
-    name: str
-    engines: Optional[List[Engine]]
